+++ internal/actions/serve/serve.go
@@ -8,8 +8,8 @@ import (
 	"os/exec"
 	"path/filepath"
 
-	"github.com/customrealms/cli/minecraft"
-	"github.com/customrealms/cli/server"
+	"github.com/customrealms/cli/internal/minecraft"
+	"github.com/customrealms/cli/internal/server"
 )
 
 type ServeAction struct {
