+++ internal/minecraft/version.go
