+++ internal/minecraft/version_papermc.go
