+++ internal/project/package_json.go
