+++ internal/server/fetcher_cache.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"path"
 
-	"github.com/customrealms/cli/minecraft"
+	"github.com/customrealms/cli/internal/minecraft"
 )
 
 type cachedFetcher struct {
