+++ internal/server/fetcher_http.go
@@ -4,7 +4,7 @@ import (
 	"io"
 	"net/http"
 
-	"github.com/customrealms/cli/minecraft"
+	"github.com/customrealms/cli/internal/minecraft"
 )
 
 type HttpFetcher struct{}
