+++ pkg/version/version.go
@@ -1,0 +1,4 @@
+package version
+
+// Version is the current version of the CLI. This is set at build time.
+var Version string
